[ui] use swipelist for ios, see #306
